Prediction Object for Survival
Source: R/PredictionSurv.R

This object stores the predictions returned by a learner of class LearnerSurv.

Format

R6Class object inheriting from Prediction.

Usage

Inherits from Prediction

# Construction
p = PredictionSurv$new(task, risk)
    # Members
p$predict_types
p$risk
p$row_ids
p$truth
    # S3 methods
as.data.table(p)

Arguments

task (Task): Task used for prediction. Used to extract row_ids and truth. Set to NULL to skip all argument checks during initialization. Slots p$row_ids and p$truth need to be set manually in this case

risk (numeric()): Vector of risk scores.

Details

$new() initializes a new object of class Prediction.

$predict_types (character) stores the predict types available: currently only "risk" is supported.

$risk stores the predicted risk scores.

row_ids stores the row IDs.

$truth stores the true survival times as survival::Surv() object.

The prediction object can be transformed to a simple data.table() with data.table::as.data.table. True survival time and status are separate columns, named "time" and "status".

Examples

library(mlr3)
task = mlr_tasks$get("lung")
learner = mlr_learners$get("surv.rpart")
e = Experiment$new(task, learner)$train()$predict()

#> INFO [mlr3] Training learner 'surv.rpart' on task 'lung' ...
#> INFO [mlr3] Predicting with model of learner 'surv.rpart' on task 'lung' ...

p = e$prediction
head(as.data.table(p))

#>    row_id      risk time status
#> 1:      1 1.8165846  306      1
#> 2:      2 1.8165846  455      1
#> 3:      3 0.6019784 1010      0
#> 4:      4 2.0298656  210      1
#> 5:      5 0.6019784  883      1
#> 6:      6 0.8886228 1022      0