Prediction Object for Survival
Source: R/PredictionSurv.R

This object stores the predictions returned by a learner of class LearnerSurv.

The task_type is set to "surv".

Format

R6::R6Class object inheriting from mlr3::Prediction.

Construction

p = PredictionSurv$new(task = NULL, risk = NULL,
  row_ids = task$row_ids, truth = task$truth())

task :: TaskClassif
Task for which the predictions are made. Used to extract the row ids and the true labels. Must be subsetted to test set.

risk :: numeric()
Vector of risk scores. One element for each observation in the test set. The higher the risk, the more likely is an event. Used in measures like mlr_measures_surv.harrells_c.

row_ids :: (integer() | character())
Row ids of the task. Per default, these are extracted from the task.

truth :: survival::Surv()
Observed survival times. Per default, these are extracted from the task.

Fields

See mlr3::Prediction.

The field task_type is set to "surv".

Examples

library(mlr3)
task = mlr_tasks$get("lung")
learner = mlr_learners$get("surv.rpart")
e = Experiment$new(task, learner)$train()$predict()

#> INFO [mlr3] Training learner 'surv.rpart' on task 'lung' ...
#> INFO [mlr3] Predicting with model of learner 'surv.rpart' on task 'lung' ...

p = e$prediction
head(as.data.table(p))

#>    row_id time status      risk
#> 1:      1  306      1 1.8165846
#> 2:      2  455      1 1.8165846
#> 3:      3 1010      1 0.6019784
#> 4:      4  210      1 2.0298656
#> 5:      5  883      1 0.6019784
#> 6:      6 1022      1 0.8886228