Survival Task
Source: R/TaskSurv.R

This task specializes mlr3::Task and mlr3::TaskSupervised for right-censored survival problems. The target column is assumed to be a factor. Predefined tasks are stored in mlr3::mlr_tasks.

The task_type is set to "surv".

Format

R6::R6Class object inheriting from Task/TaskSupervised.

Construction

t = TaskSurv$new(id, backend, time, status)

id :: character(1)
Name of the task.

backend :: DataBackend

time :: numeric()
Event times.

status :: integer()
Event indicator. "0" means alive (no event), "1" means dead (event).

Fields

backend :: DataBackend.

col_info :: data.table::data.table()
Table with with 3 columns: Column names of DataBackend are stored in columnid. Column type holds the storage type of the variables, e.g. integer, numeric or character. Column levels keeps a list of possible levels for factor and character variables.

col_roles :: named list()
Each column (feature) can have an arbitrary number of roles in the learning task:

"target": Labels to predict.

"feature": Regular feature.

"order": Data returned by data() is ordered by this column (or these columns).

"groups": During resampling, observations with the same value of the variable with role "groups" are marked as "belonging together". They will be exclusively assigned to be either in the training set or the test set for each resampling iteration. Only a single column may be marked as grouping column.

"weights": Observation weights. Only a single column may be marked as weights. col_roles keeps track of the roles with a named list of vectors of feature names. To alter the roles, use t$set_col_role().

row_roles :: named list()
Each row (observation) can have an arbitrary number of roles in the learning task:

"use": Use in train / predict / resampling.

"validation": Hold the observations back unless explicitly requested. row_roles keeps track of the roles with a named list of vectors of feature names. To alter the role, use set_row_role().

feature_names :: character()
Returns all column names with role == "feature".

feature_types :: data.table::data.table()
Returns a table with columns id and type where id are the column names of "active" features of the task and type is the storage type.

formula :: formula()
Constructs a stats::formula, e.g. [target] ~ [feature_1] + [feature_2] + ... + [feature_k], using the active features of the task.

hash :: character(1)
Hash (unique identifier) for this object.

id :: character(1)
Stores the identifier of the Task.

measures :: list() of Measure
Stores the measures to use for this task.

ncol :: integer(1)
Returns the total number of cols with role "target" or "feature".

nrow :: integer(1)
Return the total number of rows with role "use".

row_ids :: (integer() | character())
Returns the row ids of the DataBackend for observations with with role "use".

target_names :: character()
Returns all column names with role "target".

task_type :: character(1)
Stores the type of the Task.

properties :: character()
Set of task properties. Possible properties are are stored in mlr_reflections$task_properties.

groups :: data.table::data.table()
If the task has a designated column role "groups", table with two columns: "row_id" (integer() | character()) and the grouping variable group (vector()). Returns NULL if there are is no grouping column.

weights :: data.table::data.table()
If the task has a designated column role "weights", table with two columns: "row_id" (integer() | character()) and the observation weights weight (numeric()). Returns NULL if there are is no weight column.

Methods

truth(row_ids = NULL) :: data.table::data.table()
Table with true labels for specified row_ids. Defaults to all rows in use.

data(rows = NULL, cols = NULL, data_format = NULL)
(integer() | character(), character(), character()) -> any
Returns a slice of the data from the DataBackend in the data format specified by data_format (depending on the DataBackend, but usually a data.table::data.table()). It is possible to provide multiple formats, the first format supported is selected.

Rows are subsetted to only contain observations with role "use". Columns are filtered to only contain features with roles "target" and "feature". If invalid rows or cols are specified, an exception is raised.

levels(cols = NULL)
character() -> named list()
Returns the distinct values of columns in cols for columns with storage type "character", "factor" or "ordered". Argument cols defaults to all such columns with role "target" or "feature".

Note that this function ignores the row roles, it returns all levels available in the DataBackend. To update the stored level information, e.g. after filtering a task, call $droplevels().

missings(cols = NULL)
character() -> named integer()
Returns the number of missing values observations for each columns in cols. Argument cols defaults to all columns with role "target" or "feature".

head(n = 6)
integer() -> data.table::data.table()
Get the first n observations with role "use".

set_col_role(cols, new_roles, exclusive = TRUE)
(character(), character(), logical(1)) -> self
Adds the roles new_roles to columns referred to by cols. If exclusive is TRUE, the referenced columns will be removed from all other roles.

set_row_role(rows, new_roles, exclusive = TRUE)
(character(), character(), logical(1)) -> self
Adds the roles new_roles to rows referred to by rows. If exclusive is TRUE, the referenced rows will be removed from all other roles.

filter(rows)
(integer() | character()) -> self
Subsets the task, reducing it to only keep the rows specified. See the section on task mutators for more information.

select(cols)
character() -> self
Subsets the task, reducing it to only keep the columns specified. See the section on task mutators for more information.

cbind(data)
data.frame() -> self
Extends the DataBackend with additional columns. The row ids must be provided as column in data (with column name matching the primary key name of the DataBackend). If this column is missing, it is assumed that the rows are exactly in the order of t$row_ids. See the section on task mutators for more information.

rbind(data)
data.frame() -> self
Extends the DataBackend with additional rows. The new row ids must be provided as column in data. If this column is missing, new row ids are constructed automatically. See the section on task mutators for more information.

replace_features(data)
data.frame() -> self
Replaces some features of the task by constructing a completely new DataBackendDataTable. This operation is similar to calling select() and cbind(), but explicitly copies the data. See the section on task mutators for more information.

droplevels(cols = NULL)
character -> self
Updates the cache of stored factor levels, removing all levels not present in the set of active rows. cols defaults to all columns with storage type "character", "factor", or "ordered".

Examples

library(mlr3)
data("lung", package = "survival")
b = as_data_backend(lung)
task = TaskSurv$new("lung", backend = b, time = "time", status = "status")

task$target_names

#> [1] "time"   "status"

task$feature_names

#> [1] "age"       "inst"      "meal.cal"  "pat.karno" "ph.ecog"   "ph.karno" 
#> [7] "sex"       "wt.loss"  

task$formula

#> Surv(time, status) ~ age + inst + meal.cal + pat.karno + ph.ecog + 
#>     ph.karno + sex + wt.loss
#> <environment: namespace:survival>

task$truth()

#>   [1]  306   166   166   559+  450   364   107   177   156   529+   11   429 
#>  [13]  170   351    15   181   283   201   524    13   212   524   288   654 
#>  [25]  363   442   199   550    54   558   207    92    60   551+  728   543+
#>  [37]  293   202   353   511+  267   511+  371   387   457    71   337   201 
#>  [49]  404+  222    62   458+  356+  353   163    31   567   340   229   444+
#>  [61]  315+  182   156   329   364+  291   179   144   376+  384+  268   292+
#>  [73]  142   413+  266+  194   320   181   613   285   301+  348   197   382+
#>  [85]  303+  296+  180   186   145   707   269+  300+  284+  350   272+  292+
#>  [97]  332+  285   259+  110    61   286   270    81   131   225+  269   225+
#> [109]  243+  279+  276+  455    88   135    79    59   240+  202+  235+  105 
#> [121]  224+  239   237+  301   173+  252+  221+  185+   92+   13   222+  192+
#> [133]  183   211+   81   175+  197+  203+  116   188+  191+  105+  174+  177+
#> [145]  624   371   394   520   574   118   390  1010+   12   473    26   533 
#> [157]  107    53   122   814   965+   93   210   731   460   153   433   145 
#> [169]  583    95   303   519   643   883   765   735   189    53   246   689 
#> [181]   65     5   132   687  1022+  345   444   223   175    60   163    65 
#> [193]  208   821+  428   310   230   840+  305    11   132   226   426   705 
#> [205]  363    11   361   176   791    95   196+  167   806+  284   641   147 
#> [217]  740+  218   163   655   239    88   245   588+   30   179   310   477